requirements.txt
# ── Web Framework ──────────────────────────────────────────────
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Login==0.6.3

# ── OAuth2 / Google Login ──────────────────────────────────────
Authlib==1.3.2
requests==2.32.3

# ── Database ───────────────────────────────────────────────────
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9

# ── Scheduling ─────────────────────────────────────────────────
APScheduler==3.10.4

# ── Date utilities ─────────────────────────────────────────────
python-dateutil==2.9.0

# ── Environment / Config ───────────────────────────────────────
python-dotenv==1.0.1
Werkzeug==3.0.4

# ── PDF Generation ─────────────────────────────────────────────
reportlab==4.2.5

# ── QR Code (UPI) ──────────────────────────────────────────────
qrcode[pil]==7.4.2
Pillow==10.4.0

# ── Email ──────────────────────────────────────────────────────
email-validator==2.2.0

# ── Production Server ──────────────────────────────────────────
gunicorn==23.0.0
